# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Citas Médicas",
        "description": "Documentación de la API",
        "contact": {
            "email": "contacto@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/AgregarCitaMedica": {
            "post": {
                "tags": [
                    "Citas Médicas"
                ],
                "summary": "Crear una nueva cita médica",
                "operationId": "1fb2229555a424400f8561222b97a244",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CitaMedica"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cita médica creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CitaMedica"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/citaMedica/{id}": {
            "get": {
                "tags": [
                    "Citas Médicas"
                ],
                "summary": "Obtener una cita médica por ID",
                "operationId": "666b8dbc51c0f8e5ad10f8e704e786cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos de la cita médica",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CitaMedica"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cita médica no encontrada"
                    }
                }
            }
        },
        "/api/ActualizarCitaMedica/{id}": {
            "put": {
                "tags": [
                    "Citas Médicas"
                ],
                "summary": "Actualizar una cita médica",
                "operationId": "814875cd93f71f60d236e57a8e2cc550",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CitaMedica"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cita médica actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CitaMedica"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cita médica no encontrada"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/EliminarCitaMedica/{id}": {
            "delete": {
                "tags": [
                    "Citas Médicas"
                ],
                "summary": "Eliminar una cita médica",
                "operationId": "85b98c1e03c7bb1440a5596cb88c70ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cita médica eliminada con éxito"
                    },
                    "404": {
                        "description": "Cita médica no encontrada"
                    }
                }
            }
        },
        "/api/citasPorPaciente/{pacienteId}": {
            "get": {
                "tags": [
                    "Citas Médicas"
                ],
                "summary": "Obtener citas por paciente",
                "operationId": "39a2f0f810af4dc226b1d85cb8aff72d",
                "parameters": [
                    {
                        "name": "pacienteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de citas del paciente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CitaMedica"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron citas médicas para este paciente"
                    }
                }
            }
        },
        "/api/citasPorDoctor/{doctorId}": {
            "get": {
                "tags": [
                    "Citas Médicas"
                ],
                "summary": "Obtener citas por doctor",
                "operationId": "9b92486ca9f32d9d6bbfa81fb0cbef93",
                "parameters": [
                    {
                        "name": "doctorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de citas del doctor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CitaMedica"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron citas médicas para este doctor"
                    }
                }
            }
        },
        "/api/citasCompletadas": {
            "get": {
                "tags": [
                    "Citas Médicas"
                ],
                "summary": "Obtener citas completadas",
                "operationId": "e67edfdd4d342e410014e6b2f38d7329",
                "responses": {
                    "200": {
                        "description": "Lista de citas completadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CitaMedica"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron citas médicas completadas"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CitaMedica": {
                "title": "Cita Médica",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "paciente_id": {
                        "type": "integer"
                    },
                    "doctor_id": {
                        "type": "integer"
                    },
                    "fecha_hora": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "programada",
                            "completada",
                            "cancelada"
                        ]
                    },
                    "novedad": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}